+++ NEWS.md
@@ -2,6 +2,7 @@
 
 * Let users auth via an email again via `gcs_auth`
 * correct file.copy behaviour in `gcs_load_all()` (#112) - thanks @jasonmhoule
+* Allow GCS uploads if ACL has been set at bucket level (#111)
 
 # googleCloudStorageR 0.5.1